+++ vite/src/gsap.js
@@ -1,4 +1,9 @@
 import gsap from "gsap";
+import { ScrollTrigger } from "gsap/ScrollTrigger";
+import { SplitText } from "gsap/SplitText";
+
+gsap.registerPlugin(ScrollTrigger);
+gsap.registerPlugin(SplitText);
 
 const def = {
   duration: 1,
@@ -7,5 +12,7 @@ const def = {
 
 gsap.defaults(def);
 
+const utils = {};
+
 export default gsap;
-export { def };
+export { def, utils, ScrollTrigger, SplitText };