+++ vite/src/modules/dom.js
@@ -2,6 +2,13 @@ import { Text } from "./animation/text";
 // import { Track } from "../util/track";
 // import { Alpha } from "./animation/alpha";
 
+const lib = [
+  {
+    selector: '[data-a="char"],[data-a="word"],[data-a="line"]',
+    class: Text,
+  },
+];
+
 export class Dom {
   constructor() {
     this.create();
@@ -10,23 +17,33 @@ export class Dom {
   resize() {}
 
   render(t) {
-    // this.track?.render();
+    this.children.forEach((child) => {
+      if (child.render) child.render();
+    });
+  }
+
+  get _children() {
+    const arr = lib.map((item) => {
+      const { selector, class: Class } = item;
+      const elements = [...document.querySelectorAll(selector)];
+
+      if (elements.length === 0) return null;
+      return elements.map((element) => new Class(element));
+    });
+
+    return arr;
   }
 
   create() {
-    this.texts = [
-      ...document.querySelectorAll(
-        '[data-a="char"],[data-a="word"],[data-a="line"]'
-      ),
-    ].map((el) => new Text(el, {}));
+    this.children = this._children.flat().filter((child) => child !== null);
 
     this.start();
   }
 
   start() {
-    this.texts?.forEach((text) => text.start());
-    this.alpha?.start();
-    this.track?.start();
+    this.children.forEach((child) => {
+      if (child.start) child.start();
+    });
   }
 
   destroy() {