+++ vite/src/util/clientRect.js
@@ -22,7 +22,7 @@ export const clientRect = (element) => {
 
 // to check
 export const clientRectGl = (element, ratio = 1) => {
-  const bounds = element.clientRect();
+  const bounds = clientRect(element);
 
   for (const [key, value] of Object.entries(bounds))
     bounds[key] = value * ratio;